docs: split external storage snippets into individual files
Each snippet is now a separate file with its own imports so that
snipsync pulls complete, self-contained code blocks.
