+++ features/snippets/external_storage/custom_storage_driver.py
@@ -1,0 +1,59 @@
+import os
+import uuid
+from collections.abc import Sequence
+
+from temporalio.api.common.v1 import Payload
+from temporalio.converter import (
+    StorageDriver,
+    StorageDriverClaim,
+    StorageDriverStoreContext,
+    StorageDriverRetrieveContext,
+    StorageDriverWorkflowInfo,
+)
+
+
+# @@@SNIPSTART python-custom-storage-driver
+class LocalDiskStorageDriver(StorageDriver):
+    def __init__(self, store_dir: str = "/tmp/temporal-payload-store") -> None:
+        self._store_dir = store_dir
+
+    def name(self) -> str:
+        return "local-disk"
+
+    async def store(
+        self,
+        context: StorageDriverStoreContext,
+        payloads: Sequence[Payload],
+    ) -> list[StorageDriverClaim]:
+        os.makedirs(self._store_dir, exist_ok=True)
+
+        prefix = self._store_dir
+        target = context.target
+        if isinstance(target, StorageDriverWorkflowInfo) and target.id:
+            prefix = os.path.join(self._store_dir, target.namespace, target.id)
+            os.makedirs(prefix, exist_ok=True)
+
+        claims = []
+        for payload in payloads:
+            key = f"{uuid.uuid4()}.bin"
+            file_path = os.path.join(prefix, key)
+            with open(file_path, "wb") as f:
+                f.write(payload.SerializeToString())
+            claims.append(StorageDriverClaim(data={"path": file_path}))
+        return claims
+
+    async def retrieve(
+        self,
+        context: StorageDriverRetrieveContext,
+        claims: Sequence[StorageDriverClaim],
+    ) -> list[Payload]:
+        payloads = []
+        for claim in claims:
+            file_path = claim.data["path"]
+            with open(file_path, "rb") as f:
+                raw = f.read()
+            payload = Payload()
+            payload.ParseFromString(raw)
+            payloads.append(payload)
+        return payloads
+# @@@SNIPEND
